refactor: move OutputSet into util package
