+++ src/main/java/com/wmp/htmlparser/util/OutputSet.java
@@ -1,4 +1,4 @@
-package com.wmp.htmlparser.domain;
+package com.wmp.htmlparser.util;
 
 import lombok.Getter;
